distributed_map_reduce/requirements.txt
asgiref==3.7.2
Django==5.0.3
gunicorn==21.2.0
packaging==23.2
PyYAML==6.0.1
sqlparse==0.4.4
